update 2.0
now added config file for better UX
downloads will now start from the last category and pageno of the previous session.

No more Re-downloads!
added filename and size check to redownload only the half done or not downloaded files.

+++ download.py
@@ -14,6 +14,11 @@ def download(url, path):
     filename = os.path.join(path, filename)
     filesize = r.headers.get('content-length')
     filesize = int(filesize)
+
+    if os.path.exists(filename) and filesize==os.path.getsize(filename): # file exist check and previous failed download will be re-downloaded
+        print(filename,' Already Exists!')
+        return 0
+
     if r.status_code == 200:
         with open(filename, 'wb') as p3t,tqdm(unit='B', unit_scale=True, unit_divisor=1024, file=sys.stdout, total=filesize, desc=filename)as progress:
             for chunk in r.iter_content(chunk_size=chunksize):
@@ -21,7 +26,7 @@ def download(url, path):
                     p3t.write(chunk)
                     progress.update(chunksize)
     else:
-        print('NOT found', filename)
+        print(filename,'NOT Found! check the notfound file for the urls use waybackmachine to download the file')
         with open('NOTFOUND.txt', 'a') as nf:
             nf.write(str(url) + '\n')
     return 0
